refactor without observable boolean
