+++ app/src/main/java/com/infomaniak/drive/ui/bottomSheetDialogs/BackgroundSyncPermissionsViewModel.kt
@@ -17,10 +17,7 @@
  */
 package com.infomaniak.drive.ui.bottomSheetDialogs
 
-import android.app.Activity
-import android.content.Context
 import android.os.Build
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.infomaniak.drive.BuildConfig
 import kotlinx.android.synthetic.main.fragment_bottom_sheet_background_sync.*
@@ -32,24 +29,14 @@ class BackgroundSyncPermissionsViewModel : ViewModel() {
      *
      * @see hasDoneNecessaryCheckbox
      */
-    val hasDoneNecessary = MutableLiveData<Boolean>()
+    var hasDoneNecessary = false
 
     /**
      * True if the IGNORE_BATTERY_OPTIMIZATIONS permission is given to the app
      */
-    val isWhitelisted = MutableLiveData<Boolean>()
-
-    /**
-     * Requests battery permission when passed to true
-     *
-     * @see allowBackgroundSyncSwitch
-     */
-    val shouldBeWhitelisted = MutableLiveData<Boolean>()
+    var isWhitelisted = false
 
     companion object {
-
-        private const val SHARED_PREFS = "HINT_BATTERY_OPTIMIZATIONS"
-        private const val FLAG_SHOW_BATTERY_OPTIMIZATION_DIALOG = "showBatteryOptimizationDialog"
 
         /**
          * List of manufacturers which are known to restrict background processes or otherwise
@@ -67,15 +54,5 @@ class BackgroundSyncPermissionsViewModel : ViewModel() {
          * @see evilManufacturers
          */
         val manufacturerWarning = evilManufacturers.contains(Build.MANUFACTURER.lowercase()) || BuildConfig.DEBUG
-
-        fun Context.mustShowBatteryOptimizationDialog(): Boolean {
-            val sharedPrefs = getSharedPreferences(SHARED_PREFS, Activity.MODE_PRIVATE)
-            return sharedPrefs.getBoolean(FLAG_SHOW_BATTERY_OPTIMIZATION_DIALOG, true)
-        }
-
-        fun Context.updateShowBatteryOptimizationDialog(mustShowBatteryDialog: Boolean) {
-            val editableSharedPrefs = getSharedPreferences(SHARED_PREFS, Activity.MODE_PRIVATE).edit()
-            editableSharedPrefs.putBoolean(FLAG_SHOW_BATTERY_OPTIMIZATION_DIALOG, mustShowBatteryDialog).apply()
-        }
     }
 }
